Move model testing utils into designated dir
